refactor: move core to basehist
