+++ src/hist/__init__.py
@@ -10,7 +10,7 @@ from .version import version as __version__
 from . import axis
 from .general import Hist
 from .named import NamedHist
-from .core import BaseHist
+from .basehist import BaseHist
 from . import numpy
